Updated rainbow_table_gen.c
I added the number of threads to be dynamically set by the users on command line while running the binary code. the usage of the code will now be  ./ rainbow_table_gen <num_threads> e.g ./ rainbow_table_gen 6 where 6 is the number of threads used in generating the password. I also added print statement In the main function to print the following:

Number of threads used:
Size of the file generated :

Which will appear once you run the code. This will help in collecting data for analysis.